requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
pycryptodome
python-dotenv
